chore: move history clear button file colocation
