+++ src/components/HistoryDialog/HistoryClearButton.tsx
